all routes are set up including error page. layout has been added. home components are moved to home folder. navbar and foooter moved to shared folder
