+++ src/App.js
@@ -1,16 +1,11 @@
 import './App.css';
-import { Navbar } from "./components/Navbar";
-import HomePageProducts from './components/HomePageProducts/HomePageProducts';
-import Footer from './components/Footer/Footer';
-import NewsLetter from './components/NewsLetter/NewsLetter';
+import { RouterProvider } from 'react-router-dom';
+import router from './Routes/Routes/Routes';
 
 function App() {
   return (
     <div>
-      <Navbar></Navbar>
-      <HomePageProducts></HomePageProducts>
-      <NewsLetter></NewsLetter>
-      <Footer></Footer>
+      <RouterProvider router={router}></RouterProvider>
     </div>
   );
 }
