+++ src/Routes/Routes/Routes.js
@@ -1,0 +1,20 @@
+const { createBrowserRouter } = require("react-router-dom");
+const { default: ErrorPage } = require("../../components/ErrorPage/ErrorPage");
+const { default: Home } = require("../../components/Home/Home/Home");
+const { default: Main } = require("../../Layout/Main");
+
+const router = createBrowserRouter([
+    { 
+        path: "/",
+        element: <Main></Main>,
+        errorElement: <ErrorPage></ErrorPage>,
+        children: [
+            { 
+                path: "/",
+                element: <Home></Home>,
+            }
+        ]        
+    }
+]);
+
+export default router;
