+++ src/components/Home/Home/Home.js
@@ -1,0 +1,14 @@
+import React from "react";
+import HomePageProducts from "../../HomePageProducts/HomePageProducts";
+import NewsLetter from "../NewsLetter/NewsLetter";
+
+const Home = () => {
+  return (
+    <div>
+      <HomePageProducts></HomePageProducts>
+      <NewsLetter></NewsLetter>
+    </div>
+  );
+};
+
+export default Home;
