+++ src/components/Home/NewsLetter/newsLetter.css
@@ -1,6 +1,6 @@
 .bg{
   min-height: 400px;
-  background-image: url('../../../src/assets/images/newsletterbg.jpg');
+  background-image: url('../../../../src/assets/images/newsletterbg.jpg');
   background-repeat: no-repeat;
   background-position: center center;
   background-size: 100%;
